requires_new_items
git-svn-id: svn+ssh://rubyforge.org/var/svn/dectxn/trunk@32 4a9433e1-d360-4096-8ce1-a46ac5279aa0

+++ DecTxnTestApp/db/migrate/003_add_price_to_item.rb
@@ -1,6 +1,6 @@
 class AddPriceToItem < ActiveRecord::Migration
   def self.up
-    add_column :items, :price, :double
+    add_column :items, :price, :float
   end
 
   def self.down
